requirements.txt
# Binance API Client
python-binance==1.0.19

# CLI Framework
typer==0.12.3
rich==13.7.1

# Environment Variables
python-dotenv==1.0.1

# HTTP Requests
requests==2.31.0

# Utilities
pydantic==2.6.4
